refactor: move the interfaces to a folder
