+++ src/app/utils/interfaces/interface-categories.ts
