+++ src/app/utils/interfaces/interface-product-search.ts
