+++ src/app/utils/interfaces/interfaces.ts
